Refactor code structure and update imports
