+++ lib/common/bottom_nav_bar/bottom_nav_bar.dart
@@ -5,8 +5,8 @@ import 'package:ionicons/ionicons.dart';
 import 'package:water_drop_nav_bar/water_drop_nav_bar.dart';
 
 import '../../config/router/app_router.dart';
-import '../../states/widgets/bottom_nav_bar/nav_bar_logic.dart';
 import '../../utils/context_extensions.dart';
+import 'nav_bar_logic.dart';
 
 class BottomNavBar extends ConsumerWidget {
   const BottomNavBar({super.key});
@@ -26,7 +26,7 @@ class BottomNavBar extends ConsumerWidget {
         selectedIndex: nav.navIndex,
         onItemSelected: (int index) {
           ref.read(bottomNavBarLogicProvider.notifier).setNavIndex(index);
-    
+
           /// If you have more than 2 screens, you have to implement a switch case Example:
           /// case 0:
           ///  context.go(SGRoute.firstScreen.route);
