+++ lib/features/home/home.dart
@@ -6,8 +6,8 @@ import 'package:ionicons/ionicons.dart';
 
 import '../../../utils/context_extensions.dart';
 
-import '../widgets/app_bar_gone.dart';
-import '../widgets/bottom_nav_bar.dart';
+import '../../ui/widgets/app_bar_gone.dart';
+import '../../common/bottom_nav_bar/bottom_nav_bar.dart';
 import 'widgets/header.dart';
 import 'widgets/social_tile_widget.dart';
 import 'widgets/theme_widget.dart';
